removed useless cast

+++ org.eclipse.jdt.ui/core extension/org/eclipse/jdt/internal/corext/util/SuperTypeHierarchyCache.java
@@ -125,7 +125,7 @@ public class SuperTypeHierarchyCache {
 	}
 	
 	public static int getAvgCreationTime() {
-		return (int) fgTotCreationTime / fgCacheMisses;
+		return fgTotCreationTime / fgCacheMisses;
 	}			
 
 }
